pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = "Autonomous Agent for Finance"
authors = ["Jan Komisarczyk <jan.komisarczyk@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.scripts]
financebro = 'src.cli:main'

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.0"
pydantic-settings = "^2.0.3"
openai = "^0.28.1"
coloredlogs = "^15.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
